Add check HasForeignKey method to dialect, also move mssql dialect to a separate repo as it is not well tested, close #832
